requirements.txt
# Core LLM & Embeddings
openai==1.12.0
anthropic==0.25.0
sentence-transformers==2.3.1

# Vector Search & Retrieval
faiss-cpu==1.7.4
rank-bm25==0.2.2

# RAG Framework
langchain==0.1.6
langchain-community==0.0.19
langchain-openai==0.0.5

# Document Processing
pdfplumber==0.10.3
pdf2image==1.16.3
python-docx==1.1.0
Pillow==10.2.0
pytesseract==0.3.10

# NLP & Text Processing
tiktoken==0.5.2

# Data Validation
pydantic==2.6.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1
tqdm==4.66.1
requests==2.31.0

# Development & Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Phase 3: Evaluation & Quality
ragas==0.1.0

# Web API & Interface
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Optional UI
streamlit==1.30.0
